# docs/swagger.yaml
...
basePath: /api/v1
definitions:
...
  model.Element:
    properties:
      label:
        type: string
      unit:
        type: string
      value:
        type: string
    type: object
  model.ElementResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Element'
        type: array
      numItems:
        type: integer
    type: object
  model.ElementType:
    properties:
      label:
        type: string
      value:
        type: string
    type: object
  model.ElementTypeResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ElementType'
        type: array
      numItems:
        type: integer
    type: object
...
  model.GeologicalSetting:
    properties:
      setting:
        type: string
    type: object
  model.GeologicalSettingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.GeologicalSetting'
        type: array
      numItems:
        type: integer
    type: object
...
  model.Location:
    properties:
      name:
        type: string
    type: object
  model.LocationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Location'
        type: array
      numItems:
        type: integer
    type: object
...
  model.Site:
    properties:
      latitude:
        type: number
      locationPrecision:
        type: number
      locationPrecisionComment:
        type: string
      longitude:
        type: number
      samplingFeatureID:
        type: integer
      setting:
        type: string
      siteDescription:
        type: string
      siteTypeCV:
        type: string
      spatialReferenceID:
        type: integer
    type: object
  model.SiteResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Site'
        type: array
      numItems:
        type: integer
    type: object
...
host: api-test.georoc.eu
info:
  contact:
    email: digis-info@uni-goettingen.de
    name: DIGIS Project
    url: https://www.uni-goettingen.de/de/643369.html
  description: |-
    This is the database api for the new GeoROC datamodel

    Note: Semicolon (;) in queries are not allowed and need to be url-encoded as per this issue: golang.org/issue/25192
  license:
    name: Data retrieved is licensed under CC BY-SA 4.0
    url: https://creativecommons.org/licenses/by-sa/4.0/
  title: DIGIS Database API
  version: 0.4.1
paths:
...
  /queries/locations/l1:
    get:
      consumes:
      - application/json
      description: get top level locations
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LocationResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve locations of first level
      tags:
      - locations
  /queries/locations/l2:
    get:
      consumes:
      - application/json
      description: get second level locations
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      - description: Locationname Level 1
        in: query
        name: locationl1
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LocationResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve locations of second level
      tags:
      - locations
  /queries/locations/l3:
    get:
      consumes:
      - application/json
      description: get third level locations
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      - description: Locationname Level 1
        in: query
        name: locationl1
        required: true
        type: string
      - description: Locationname Level 2
        in: query
        name: locationl2
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LocationResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve locations of third level
      tags:
      - locations
  /queries/results/elements:
    get:
      consumes:
      - application/json
      description: get chemical elements
      parameters:
      - description: Element type
        in: query
        name: type
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ElementResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve chemical elements
      tags:
      - samples
  /queries/results/elementtypes:
    get:
      consumes:
      - application/json
      description: get chemical element types
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ElementTypeResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve chemical element types
      tags:
      - samples
...
  /queries/sites:
    get:
      consumes:
      - application/json
      description: Get all sites
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SiteResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve all sites
      tags:
      - sites
  /queries/sites/{samplingfeatureID}:
    get:
      consumes:
      - application/json
      description: Get sites by samplingfeatureID
      parameters:
      - description: samplingfeatureID
        in: path
        name: samplingfeatureID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SiteResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve sites by samplingfeatureID
      tags:
      - sites
  /queries/sites/settings:
    get:
      consumes:
      - application/json
      description: Get all geological settings
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeologicalSettingResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve all geological settings
      tags:
      - sites
...
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    description: Accesskey based security scheme to secure api groups "/queries/*"
      and "/geodata/*"
    in: header
    name: DIGIS-API-ACCESSKEY
    type: apiKey
swagger: "2.0"